requirements/local.txt
-r base.txt

django-suit==0.2.12
